Delete study steps, move steps panel
