--- src/admin/StudyStep.tsx
@@ -1,3 +1,0 @@
-export function StudyStep(props: { stepId: string }) {
-  return <div>Step {props.stepId}</div>;
-}
